+++ Runtime/Scripts/Tezos/ITezos.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using TezosSDK.Tezos.API;
 using TezosSDK.Tezos.API.Models;
+using TezosSDK.Tezos.API.Models.Abstract;
 using TezosSDK.Tezos.Wallet;
 
 namespace TezosSDK.Tezos
@@ -22,7 +23,7 @@ namespace TezosSDK.Tezos
         /// <summary>
         /// Currently used FA2 Contract.
         /// </summary>
-        TokenContract TokenContract { get; }
+        IFA2 TokenContract { get; }
 
         /// <summary>
         /// Current wallet tz balance.