+++ Runtime/Scripts/Tezos/Tezos.cs
@@ -5,6 +5,7 @@ using TezosSDK.Beacon;
 using TezosSDK.Tezos.API;
 using TezosSDK.Tezos.API.Models;
 using TezosSDK.Tezos.API.Models.Filters;
+using TezosSDK.Tezos.API.Models.Abstract;
 using TezosSDK.Tezos.Wallet;
 using UnityEngine;
 
@@ -19,7 +20,7 @@ namespace TezosSDK.Tezos
         public WalletMessageReceiver MessageReceiver { get; }
         public ITezosAPI API { get; }
         public IWalletProvider Wallet { get; }
-        public TokenContract TokenContract { get; set; }
+        public IFA2 TokenContract { get; set; }
 
         public Tezos()
         {