+++ Runtime/Scripts/Tezos/TezosSingleton.cs
@@ -6,6 +6,7 @@ using TezosSDK.DesignPattern.Singleton;
 using TezosSDK.Tezos.API;
 using TezosSDK.Tezos.API.Models;
 using TezosSDK.Tezos.API.Models.Filters;
+using TezosSDK.Tezos.API.Models.Abstract;
 using TezosSDK.Tezos.Wallet;
 using UnityEngine;
 using Logger = TezosSDK.Helpers.Logger;
@@ -16,9 +17,9 @@ namespace TezosSDK.Tezos
     public class TezosSingleton : SingletonMonoBehaviour<TezosSingleton>, ITezos
     {
         private static Tezos _tezos;
-        public ITezosAPI API => _tezos.API;
-        public IWalletProvider Wallet => _tezos.Wallet;
-        public TokenContract TokenContract => _tezos.TokenContract;
+        public ITezosAPI API => _tezos?.API;
+        public IWalletProvider Wallet => _tezos?.Wallet;
+        public IFA2 TokenContract => _tezos?.TokenContract;
 
         protected override void Awake()
         {
@@ -27,6 +28,14 @@ namespace TezosSDK.Tezos
             Logger.CurrentLogLevel = Logger.LogLevel.Debug;
             TezosConfig.Instance.Network = NetworkType.ghostnet;
             _tezos = new Tezos();
+        }
+        
+        void OnApplicationQuit()
+        {
+            if (Wallet is IDisposable disposable)
+            {
+                disposable.Dispose();
+            }
         }
 
         public IEnumerator GetCurrentWalletBalance(Action<ulong> callback)