+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [unreleased]
+## [Unreleased]
+### Fixed
+- [Bug](https://github.com/trilitech/tezos-unity-sdk/issues/102) with standalone IL2CPP build fails
+- [Bug](https://github.com/trilitech/tezos-unity-sdk/issues/103) with closing Beacon database connections before app exit
+
+
+## [1.5.1] - 2023-06-27
 ### Fixed
 - [Bug](https://github.com/trilitech/tezos-unity-sdk/issues/99) with game items transfer not working
 
@@ -107,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added auto releases with GH actions
 
 
-[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/1.5.0...HEAD
+[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/1.5.1...HEAD
+[1.5.1]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.5.1
 [1.5.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.5.0
 [1.4.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.4.0
 [1.3.1]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.3.1